% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fact_Ajus.R
\name{fa}
\alias{fa}
\title{Factores de ajuste.}
\usage{
fa(data, d = 12)
}
\arguments{
\item{data}{base de datos suavizados con la funcion del promedio movil1 y promedio movil2.}

\item{d}{(vector) variable que determina el orden del promedio movil.}
}
\value{
Un data frame con los cálculos de los factores de ajsute.
}
\description{
Realiza calculos de ajuste en los datos suavizados.
}
\examples{
\dontrun{
# Recibe los datos calculados por la segunda función
data <- fa(data, d = 12)
}
}
