% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grafico.R
\name{grafica}
\alias{grafica}
\title{Gráfico del pronostico del periodo t, más el pronóstico al futuro.}
\usage{
grafica(x, data, di = 12)
}
\arguments{
\item{x}{(vector) es la variable independiente.
@param d (vector) variable que determina la distancia de la predicción a ralizar.}

\item{base}{(data.frame) base de datos modificados con las funciones anteriores.}
}
\value{
Devuelve una grafica del pronostico.
}
\description{
Realiza el pronóstico a la distancia requerida.
}
\examples{
\dontrun{
# Recibe los factores de ajustes calculados por la tercera función para hacer predicciones al futuro.
grafica(x="produccion", data, di = 12)
}
}
