% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pro_Mov1.R
\name{promedio_movil1}
\alias{promedio_movil1}
\title{Promedio Móvil1.}
\usage{
promedio_movil1(x, data, d = 12)
}
\arguments{
\item{x}{(vector) es la variable independiente.}

\item{data}{(data.frame) base de datos.}

\item{d}{(vector) variable que determina el orden del promedio movil a trabajar.}
}
\value{
Un data frame con los primeros datos suavizados.
}
\description{
Realiza cálculos de suavizamiento de los datos.
}
\examples{
\dontrun{
# Directorio de trabajo
ruta <-"C:/Users/GREGORIO/Desktop/Manual del Paquete DPM/Datos de producción de leche del estado de Zacatecas/base de datos.xlsx"
data <- openxlsx::read.xlsx(ruta)

# Esforzar a la base de datos que sea data.frame
data <- as.data.frame(data)

# Cargar la libreria
library(PMD)
data <- promedio_movil1(x="produccion", data, d = 12)
}
}
